Extracted all of the code from the ConventionHelper class into the corresponding class
