+++ src/DotNetToolkit.Repository/Helpers/InMemoryCache.cs
@@ -7,8 +7,14 @@
 
     internal class InMemoryCache
     {
+        #region Fields
+
         private static volatile InMemoryCache _instance;
         private static readonly object _syncRoot = new object();
+
+        #endregion
+
+        #region Properties
 
         public static InMemoryCache Instance
         {
@@ -31,11 +37,17 @@
         public ConcurrentDictionary<Tuple<Type, Type>, IEnumerable<PropertyInfo>> ForeignKeyMapping { get; }
         public ConcurrentDictionary<Type, string> TableNameMapping { get; }
 
+        #endregion
+
+        #region Constructors
+
         private InMemoryCache()
         {
             PrimaryKeyMapping = new ConcurrentDictionary<Type, IEnumerable<PropertyInfo>>();
             ForeignKeyMapping = new ConcurrentDictionary<Tuple<Type, Type>, IEnumerable<PropertyInfo>>();
             TableNameMapping = new ConcurrentDictionary<Type, string>();
         }
+
+        #endregion
     }
 }
